backend/requirements.txt
uvicorn
fastapi 
ruamel.yaml
easydict
loguru
psycopg2
SQLAlchemy
SQLAlchemy-Utils
python-dotenv
bcrypt
pytz
PyYAML
PyJWT
python-multipart